+++ src/_pytest/warning_types.py
@@ -49,12 +49,6 @@ class PytestDeprecationWarning(PytestWarning, DeprecationWarning):
     __module__ = "pytest"
 
 
-class PytestRemovedIn8Warning(PytestDeprecationWarning):
-    """Warning class for features that will be removed in pytest 8."""
-
-    __module__ = "pytest"
-
-
 class PytestRemovedIn9Warning(PytestDeprecationWarning):
     """Warning class for features that will be removed in pytest 9."""