+++ src/pytest/__init__.py
@@ -1,7 +1,5 @@
 # PYTHON_ARGCOMPLETE_OK
 """pytest: unit and functional testing with Python."""
-from typing import TYPE_CHECKING
-
 from _pytest import __version__
 from _pytest import version_tuple
 from _pytest._code import ExceptionInfo
@@ -75,7 +73,6 @@ from _pytest.warning_types import PytestCollectionWarning
 from _pytest.warning_types import PytestConfigWarning
 from _pytest.warning_types import PytestDeprecationWarning
 from _pytest.warning_types import PytestExperimentalApiWarning
-from _pytest.warning_types import PytestRemovedIn8Warning
 from _pytest.warning_types import PytestRemovedIn9Warning
 from _pytest.warning_types import PytestReturnNotNoneWarning
 from _pytest.warning_types import PytestUnhandledCoroutineWarning
@@ -139,7 +136,6 @@ __all__ = [
     "PytestConfigWarning",
     "PytestDeprecationWarning",
     "PytestExperimentalApiWarning",
-    "PytestRemovedIn8Warning",
     "PytestRemovedIn9Warning",
     "PytestReturnNotNoneWarning",
     "Pytester",
@@ -170,13 +166,3 @@ __all__ = [
     "xfail",
     "yield_fixture",
 ]
-
-if not TYPE_CHECKING:
-
-    def __getattr__(name: str) -> object:
-        if name == "Instance":
-            # The import emits a deprecation warning.
-            from _pytest.python import Instance
-
-            return Instance
-        raise AttributeError(f"module {__name__} has no attribute {name}")